Error message added to apstraClass
