+++ apstraGo/apstraGoYaml.py
@@ -43,7 +43,7 @@ def loadYaml(filepath: str) -> dict:
         try:
             inputDict=yaml.safe_load(stream)
         except yaml.YAMLError as e:
-            customError(errorInfo=e)
+            self.customError(errorInfo=e)
             exit(1)
     return inputDict
 
@@ -59,7 +59,7 @@ def yamlValidate(inputDict: dict, yamlValidateFilename: str) -> None:
     try:
         schema = eval(open(yamlValidateFilename, 'r').read())
     except (ValueError, SyntaxError) as e:
-        customError(response=e)
+        self.customError(response=e)
         exit(1)
     
     #Compared input with allowed values
@@ -67,12 +67,12 @@ def yamlValidate(inputDict: dict, yamlValidateFilename: str) -> None:
         v = Validator(schema)
         v.validate(inputDict, schema)
     except Exception as e:
-        customError(response=e)
+        self.customError(response=e)
         exit(1)
 
     #Print and exit on error
     if bool(v.errors) is not False:
-        customError(response=v.errors)
+        self.customError(response=v.errors)
         exit(1)
 
 def login(inputDict: dict) -> bytes:
@@ -86,17 +86,17 @@ def login(inputDict: dict) -> bytes:
     """    
     
     if "login" not in inputDict['apstrago']:
-        customError(response="\nLogin Section Missing From YAML File\n")
+        self.customError(response="\nLogin Section Missing From YAML File\n")
         exit(1)    
     if "username" not in inputDict['apstrago']['login']:
-        customError(response="\nUsername Missing From Login Section In YAML File\n")
+        self.customError(response="\nUsername Missing From Login Section In YAML File\n")
         exit(1)
     if "password" not in inputDict['apstrago']['login']:
-        customError(response="\nPassword Missing From Login Section In YAML File\n")
+        self.customError(response="\nPassword Missing From Login Section In YAML File\n")
         print('error')
         exit(1)
     if "address" not in inputDict['apstrago']['login']:
-        customError(response="\nAddress Missing From Login Section In YAML File\n")
+        self.customError(response="\nAddress Missing From Login Section In YAML File\n")
         exit(1)
 
     con = jcl()
@@ -107,10 +107,10 @@ def login(inputDict: dict) -> bytes:
                 customerName=inputDict['apstrago'].get('customerName'))
 
     if 'errors' in response.json():
-        customError(response)
+        self.customError(response)
         exit(1)
     else:
-        customSuccess(response)
+        self.customSuccess(response)
         return con
 
 def inputProcess(inputDict: dict, con: bytes) -> None:
@@ -149,27 +149,27 @@ def resourcePools(inputs: dict, con: bytes) -> None:
                     lastAsn=item['lastASN'])
 
             if 'errors' in response.json():
-                customError(response)
+                self.customError(response)
             else:
-                customSuccess(response)
+                self.customSuccess(response)
 
     if 'ipPools' in inputs:
         for item in inputs['ipPools']:
             response = con.resourceIpCreate(poolName=item['poolName'], network=item['network'])
 
             if 'errors' in response.json():
-                customError(response)
+                self.customError(response)
             else:
-                customSuccess(response)
+                self.customSuccess(response)
 
     if 'vniPools' in inputs:
         for item in inputs['vniPools']:
             response = con.resourceVniCreate(poolName=item['poolName'], firstVni=item['firstVNI'], lastVni=item['lastVNI'])
 
             if 'errors' in response.json():
-                customError(response)
+                self.customError(response)
             else:
-                customSuccess(response)
+                self.customSuccess(response)
 
 def rackCreation(inputs: dict, con:bytes) -> None:
     """Creates racks if config present
@@ -189,9 +189,9 @@ def rackCreation(inputs: dict, con:bytes) -> None:
                         LeafServerLinkSpeedValue=item.get('LeafServerLinkSpeedValue') or None, linksPerLeafCount=item.get('linksPerLeafCount') or None)
 
         if 'errors' in response.json():
-            customError(response)
+            self.customError(response)
         else:
-            customSuccess(response)
+            self.customSuccess(response)
 
 def templateCreation(inputs: dict, con: bytes) -> None:
     """Creates  DC templates if correct config present
@@ -206,9 +206,9 @@ def templateCreation(inputs: dict, con: bytes) -> None:
                         asnAllocation=item.get('asnAllocation') or None, overlayControl=item.get('overlayControl') or None)
         
         if 'errors' in response.json():
-            customError(response)
+            self.customError(response)
         else:
-            customSuccess(response)
+            self.customSuccess(response)
 
 def blueprintCreation(inputs: dict, con: bytes) -> None:
     """Creates blueprints if correct config present
@@ -221,9 +221,9 @@ def blueprintCreation(inputs: dict, con: bytes) -> None:
         response = con.blueprintCreate(blueprintName=item['name'], templateName=item['templateName'])
 
         if 'errors' in response.json():
-            customError(response)
+            self.customError(response)
         else:
-            customSuccess(response)
+            self.customSuccess(response)
 
 def addDevices(inputs: dict, con: bytes) -> None:
     """Creates devices if correct config present
@@ -238,15 +238,15 @@ def addDevices(inputs: dict, con: bytes) -> None:
 
     for response in responseAll:
         if 'errors' in response.json():
-            customError(response)
+            self.customError(response)
         else:
-            customSuccess(response)
+            self.customSuccess(response)
 
     if 'acknowledge' in inputs:
         if inputs['acknowledge'] == True:
             time.sleep(18)
             responseAllAck = con.ackManagedDevicesAll()
-            customSuccess(response="\n     All Devices Acknowledged\n")
+            self.customSuccess(response="\n     All Devices Acknowledged\n")
 
 def cli(argv: bytes=None) -> dict:
     """Args Parse for user input
@@ -264,36 +264,6 @@ def cli(argv: bytes=None) -> dict:
 
     args = parser.parse_args(argv)
     return args
-
-def customError(response: bytes) -> None:
-    """Print Error to console
-
-    Args:
-        response (bytes): Error Data
-    """    
-    print("#" * 120)
-    if isinstance(response, requests.models.Response):
-        print(f"""\n[bold red]Error[/]:\n     {str(response.json()['errors'])} \n     {response.reason} \n     {response.text} \n     {response.url} \n     {response.status_code}\n""")
-    elif isinstance(response, str):
-        print(f'''[bold red]Error[/]:\n     {response}''')
-    elif isinstance(response, dict):
-        print(f'''\n[bold red]Error: YAML Input Issue[/]\n     {json.dumps(response, indent=4)}\n''')
-    print("#" * 120)
-    print('\n')
-
-def customSuccess(response: bytes) -> None:
-    """Print msg to console
-
-    Args:
-        response (bytes): Msg Data
-    """    
-    print("#" * 120)
-    if isinstance(response, requests.models.Response):
-        print(f"""\n[bold green]Success[/]:\n     {response.reason} \n     {response.text} \n     {response.url} \n     {response.status_code}\n""")
-    elif isinstance(response, str):
-        print(f"""\n[bold green]Success[/]:\n     {response}""")    
-    print("#" * 120)
-    print('\n')
 
 def main():
     """Main callable function
