+++ apstraGo/jclClass.py
@@ -1,5 +1,3 @@
-from rich import print
-# from json import JSONDecodeError
 import json
 try:
     import apstraClass
@@ -185,26 +183,6 @@ class jcl(apstraClass.apstra):
                 self.customSuccess(response)
         except json.JSONDecodeError as e:
             self.customError(response=e)
-
-    def customError(self, response: bytes) -> print:
-        print("#" * 120)
-        if isinstance(response, requests.models.Response):
-            print(f"""\n[bold red]Error[/]:\n     {str(response.json()['errors'])} \n     {response.reason} \n     {response.text} \n     {response.url} \n     {response.status_code}\n""")
-        elif isinstance(response, str):
-            print(f'''[bold red]Error[/]:\n     {response}''')
-        elif isinstance(response, dict):
-            print(f'''\n[bold red]Error: YAML Input Issue[/]\n     {json.dumps(response, indent=4)}\n''')
-        print("#" * 120)
-        print('\n')
-
-    def customSuccess(self, response: bytes) -> print:
-        print("#" * 120)
-        if isinstance(response, requests.models.Response):
-            print(f"""\n[bold green]Success[/]:\n     {response.reason} \n     {response.text} \n     {response.url} \n     {response.status_code}\n""")
-        elif isinstance(response, str):
-            print(f"""\n[bold green]Success[/]:\n     {response}""")    
-        print("#" * 120)
-        print('\n')
 
     def jclProvision(self, customerName: str) -> None:
         """Create the entire JCL testbed in one easy python call.
